requirements.txt
crewai[tools]==0.22.4
python-dotenv==1.0.0